pyproject.toml
[project]
name = "job-tracker"
version = "0.1.0"
description = "Automated job application tracker that reads Gmail and updates Google Sheets with multilingual support (English/German)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.115.0",
    "gspread>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "python-dateutil>=2.8.2",
    "rapidfuzz>=3.6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "click>=8.1.7",
]
